Minor changes from int to size_t, in the course of debugging #138
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@915 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/Detection/ObjectGrower.cc
@@ -157,7 +157,7 @@ namespace duchamp {
     long xpt=vox.getX();
     long ypt=vox.getY();
     long zpt=vox.getZ();
-    long spatsize=this->itsArrayDim[0]*this->itsArrayDim[1];
+    size_t spatsize=this->itsArrayDim[0]*this->itsArrayDim[1];
     long zero = 0;
     // std::cerr << "--> " << xpt << " " << ypt << " " << zpt << "\n";
 
@@ -170,7 +170,7 @@ namespace duchamp {
       
     // std::cerr << xmin << " " << xmax << "  " << ymin << " " << ymax << "  " << zmin << " " << zmax << "\n";
     //loop over surrounding pixels.
-    int pos;
+    size_t pos;
     Voxel nvox;
     std::vector<Voxel> morevox;
     for(int x=xmin; x<=xmax; x++){
